add cast to rep macros

+++ template/template.hpp
@@ -49,10 +49,10 @@ template <class T, class U> using dict = hash_map<T, U>;
 
 #endif
 
-#define rep(i, a, b) for (ll i = (a); i < ll(b); i++)
-#define REP(i, a, b) for (ll i = (a); i <= ll(b); i++)
-#define rrep(i, a, b) for (ll i = ll(b) - 1; i >= (a); i--)
-#define RREP(i, a, b) for (ll i = ll(b); i >= (a); i--)
+#define rep(i, a, b) for (ll i = ll(a); i < ll(b); i++)
+#define REP(i, a, b) for (ll i = ll(a); i <= ll(b); i++)
+#define rrep(i, a, b) for (ll i = ll(b) - 1; i >= ll(a); i--)
+#define RREP(i, a, b) for (ll i = ll(b); i >= ll(a); i--)
 #define fore(x, a) for (auto&& x : (a))
 #define all(a) (a).begin(), (a).end()
 #define rall(a) (a).rbegin(), (a).rend()
